--- src/basic/shared/libs/price.ts
@@ -1,14 +1,0 @@
-// const formatPrice = (price: number, productId?: string): string => {
-//   if (productId) {
-//     const product = products.find((p) => p.id === productId);
-//     if (product && getRemainingStock(product) <= 0) {
-//       return "SOLD OUT";
-//     }
-//   }
-
-//   if (isAdmin) {
-//     return `${price.toLocaleString()}원`;
-//   }
-
-//   return `₩${price.toLocaleString()}`;
-// };
